pyproject.toml
[project]
name = "me3-manager"
version = "1.1.6"
description = "GUI application for Mod engine 3"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
authors = [
  { name = "2Pz" },
]
dependencies = [
  "pyside6-essentials>=6.9.2",
  "requests>=2.32.5",
  "tomli-w>=1.2.0",
  "tomlkit>=0.13.0",
]

[project.scripts]
me3-manager = "me3_manager.main:main"

[build-system]
requires = ["uv_build>=0.8.8,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "cx-freeze>=8.4.0 ; sys_platform == 'win32'",
  "pyinstaller>=6.15.0 ; sys_platform == 'linux'",
  "ruff>=0.12.12",
]

[tool.uv]
# Don't install dev dependencies on 'uv run'
default-groups = []

[tool.ruff.lint]
select = [
  "E4",
  "E7",
  "E9",
  "F",
  "I", # isort
  "B", # flake8-bugbear
  "LOG", # flake8-logging
  "G", # flake8-logging-format
  "PIE", # flake8-pie
  "T20", # flake8-print
  "UP", # pyupgrade
]
ignore = [
  # Allow `pass` (more obvious that it's a placeholder)
  "PIE790",
  # Allow redundant mode="r" arg for open()
  "UP015",
]

[tool.ruff.lint.per-file-ignores]
# Ok to use print() in standalone scripts
"scripts/*" = ["T20"]
